Broke out submitTracklist function to be imported into main.js using browserify. #2
